requirements.txt
requests>=2.28.0
pandas>=1.5.0
scikit-learn>=1.2.0
streamlit>=1.20.0
joblib>=1.2.0
python-dotenv>=0.21.0
plotly>=5.10.0
seaborn>=0.12.0
matplotlib>=3.6.0
numpy>=1.23.0